+++ minecraft_mod_manager/__main__.py
@@ -1,22 +1,14 @@
-from selenium.webdriver.chrome.webdriver import WebDriver
-from selenium.webdriver.remote.remote_connection import LOGGER
-from .mod_not_found_exception import ModNotFoundException
-from .curse_api import CurseApi
+from typing import List
+from .mod import Mod, RepoTypes
 from .dir_parser import DirParser
 from .config import config
 from .db import Db
-from .mod import Mod, RepoTypes
-from .downloader import Downloader
-from . import web_driver
-from .logger import LogColors, Logger
-from typing import List
-from os import remove, path
-import logging
+from .updater import Updater
+from .configurer import Configurer
+from .logger import Logger, LogColors
 
 
 def main():
-    webdriver = web_driver.get()
-    LOGGER.setLevel(logging.WARNING)
     db = Db()
     try:
         # Get mods in dir and sync with DB
@@ -26,68 +18,52 @@ def main():
 
         # ACTION: Update
         if config.action == "update":
-            _update(webdriver, db, installed_mods)
+            updater = Updater(db)
+            try:
+                updater.update(installed_mods)
+            finally:
+                updater.close()
 
         elif config.action == "install":
             # TODO install
             pass
 
         elif config.action == "configure":
-            # TODO configure
+            configurer = Configurer(db)
+            configurer.configure(installed_mods)
+
+        elif config.action == "list":
+            _list_mods(installed_mods)
             pass
+    except:
+        if db:
+            db.close()
 
-    finally:
-        webdriver.close()
-        db.close()
 
+def _list_mods(installed_mods: List[Mod]):
+    mod_id_max_length = 0
+    mod_repo_name_max_length = 0
 
-def _update(webdriver: WebDriver, db: Db, installed_mods: List[Mod]):
-    # Remove mods that hasn't been supplied with args
-    # But use all if no mods were supplied
-    mods_to_update = installed_mods.copy()
-    if len(config.mods) > 0:
-        mods_to_update: List[Mod] = []
-        for mod in installed_mods:
-            if mod.repo_name_alias in config.mods or mod.id in config.mods:
-                mods_to_update.append(mod)
+    for mod in installed_mods:
+        if len(mod.id) > mod_id_max_length:
+            mod_id_max_length = len(mod.id)
+        if len(mod.repo_name_alias) > mod_repo_name_max_length:
+            mod_repo_name_max_length = len(mod.repo_name_alias)
 
-    curse_parser = CurseApi(webdriver)
-    downloader = Downloader()
-    for mod in mods_to_update:
-        latest_version = None
-        was_unknown = mod.repo_type == RepoTypes.unknown
+    padding = 4
+    mod_id_width = mod_id_max_length + padding
+    mod_repo_alias_width = mod_repo_name_max_length + padding
 
-        # Curse
-        try:
-            if mod.repo_type == RepoTypes.curse or mod.repo_type == RepoTypes.unknown:
-                latest_version = curse_parser.get_latest_version(mod)
-        except ModNotFoundException as exception:
-            Logger.error(str(exception), exit=True)
+    message = f"{LogColors.bold}Installed mods:{LogColors.no_color}\n"
+    message += f"Mod".ljust(mod_id_width) + "Alias".ljust(mod_repo_alias_width) + "Site"
+    for mod in installed_mods:
+        message += "\n" + f"{mod.id}".ljust(mod_id_width)
+        message += f"{mod.repo_name_alias}".ljust(mod_repo_alias_width)
 
-        # Update DB mod if a repo was found with a matching mod name
-        if was_unknown and mod.repo_type != RepoTypes.unknown:
-            db.update_mod(mod)
+        if mod.repo_type != RepoTypes.unknown:
+            message += f"{mod.repo_type.value}"
 
-        if latest_version:
-            # Only download if not same
-            if mod.file != latest_version.filename:
-                downloaded_file = downloader.download(mod, latest_version)
-            else:
-                downloaded_file = "skip"
-
-            if downloaded_file:
-                # Remove old file
-                if downloaded_file != "skip":
-                    Logger.info(
-                        f"Updated {mod.repo_name_alias} ➡  {latest_version.name}",
-                        LogColors.green,
-                    )
-                    remove(path.join(config.dir, mod.file))
-                    mod.file = downloaded_file
-
-                # Update DB
-                mod.upload_time = latest_version.upload_time
-                db.update_mod(mod)
+    Logger.info(message)
 
 
 if __name__ == "__main__":